# Fix Performance Issue on Async Responses