--- a/ring-servlet/src/ring/util/servlet.clj
+++ b/ring-servlet/src/ring/util/servlet.clj
@@ -84,6 +84,8 @@
     (if (nil? context)
       os
       (proxy [java.io.FilterOutputStream] [os]
+        (write [b off len]
+          (.write os b off len))
         (close []
           (.close os)
           (.complete context))))))