@weavejester 2021-01-13
Thanks for the PR, and apologies for taking a little while to look at this in detail.

The code looks good, and the fix is very much appreciated. Can you change the commit message to:

```
Fix performance issue with async responses

For async responses, the function rest.util.servlet/make-output-stream
creates a subclass of java.io.FilterOutputStream in order to complete
the async context after the output stream has been closed.

Unfortunately for the performance, FilterOutputStream implements the
write(byte[], int, int) method by looping through the byte array and
calling the single byte write(int) method.

The solution is simple: add an appropriate write method to the proxy,
that calls write(byte[], int, int) on the underlying servlet output
stream.
```

The details of the full investigation can be left to the PR if anyone's interested in looking it up. The subject line of commits by convention don't capitalise every word.